requirement.txt
argh==0.26.1
cairocffi==0.7.2
CairoSVG==1.0.16
cffi==1.1.2
cssselect==0.9.1
django-wkhtmltopdf==2.0.3
docutils==0.12
lockfile==0.10.2
lxml==3.4.4
pathtools==0.1.2
pdfkit==0.5.0
pycparser==2.14
pygal==2.0.1
python-daemon==2.0.5
python-dateutil==2.4.2
PyYAML==3.11
six==1.9.0
tinycss==0.3
watchdog==0.8.3
xmltodict==0.9.2
